REWORK CMAKELIST TO BECOME FILTER SPECIFIC
The previous CMakeList was creating static libs and linking them after
in a non trivial way.
This was creating some issues when recompiling a modified file.
The static lib was updated but not the final binary.

Now the main CMakeList is only in charge of calling the CMakeList
of each filter it needs to compile.

Compile time is longer at fisrt but much quicker during development.

CMakeList are easier to read and no longer require useless dependencies.

The behavior stays the same.