ADDED THREADGROUP AND GENERATOR ABSTRACT
All generators were sharing the same methods but with duplicate code.
Now the generators inherit from an abstract implementing the common code

We now use the main process ressources to process requests.
The main process is considered to be part of the processing threads.

On stopping signal, no longer clean the sessions before the handlers
are finished running.
This removes memory errors, leaks, bad_weak_ptr and bad file descriptors

Now stopping io_context to prevent any callback push in the queue
during a graceful exit.
Added it to monitoring thread too.

Getting back to accepting sessions one by one
because of strange behaviour.
Still WIP.

Applied the generator changes to the log filter.

Added a mutex to the log filter because of undefined behaviour
using filestreams in a threaded environment.