WORKING ALERTMANAGER - HAND TESTED
Core:
AlertManager is now fully usable.
The Core method for loading the associated configuration are imlemented.
The new configuration fields are:
 - 'log_file_path': path to the log file dedicated to alerts
 - 'redis_socket_path': path to the redis UNIX socket to send alert to
 - 'alert_redis_list_name': name of the alert list in redis
 - 'alert_redis_channel_name': name of redis channel to publish alert to

If the alert configuration is missing or wrong,
it is not applied.
If the alert configuration is partialy correct,
apply as much as possible.
In both cases a warning is logged.

DARWIN_RAISE_ALERT(str) macro makes filter code easier to read.
(AlertManager.hpp)

CMake:
Added the redis dependencie to the CMakeLists.txt because it will soon
be part of all the filters through the Alert Manager.

Added the required toolkit parts to the core sources.

Added DARWIN_LIBRARIES var to remove duplicata and
avoid core libraries mistake.

fhostlookup:
Migration to the Alert Manager done and hand tested.