WIP: CONFIG UPDATE & SCRIPT LAUNCHING
Deleted useless example files.

Updated example config file.

Added proper script lauching function.

Defined a sending method to send back results from the scirpt.
Nedded to redifine a method because the Session's one is too restrictive
and the format is quite unique.
We cancel any standard result sendig for this filter
as it is for a particular usage.