USABLE SOFA IMPLEMENTATION
Fix: finspection cmake include dir

Sofa algorithm implementation is now functionnal
and provided with a default configuration file.

Because of the unusual nature of the data returned by the algorithm it needs to
override the Session's method for sending back response to the client.

This algorithm is not open-source and will be used in  a unique
setup requirering the non generic development.

Because the python code called by the filter has to keep all the data in
memory we write it on disk so we free up our memory space.
This filter is not made for realtime so optimization of memory is
preferred.