WIP: REMOVING FLOGS AND UPDATING TESTS
Flogs was not in the darwin philosophy because it was not intended to analyse
data provided for security purposes.
Moreover it was not performance focused to add another socket communication
for no further processing.

Removed flogs sources, tests and configuration files.

Updating tests that were using flogs to use ftest instead.