REMOVED TANOMALY LEGACY CONFIG
As we remove flogs we also remove legacy options of ftanomaly
`redis_list_name` & `redis_channel_name`.
We also remove the duplicate parsing of the configuration from the filter
to let it only in the AlertManager.

Removed legacy configuration tests for tanomaly.

Added a "configuration fail" trigger to ftest for test purposes.

Updated Manager tests to use ftest instead of flogs.