SESSION CLIENT RESPONSE IMPROVEMENTS
Fix: In case of parsing error of the body with a client waiting for
response would not respond and go back to listening.
Now send a json with an error message and an error code.

Now the Session::SendToClient method can send a body.
The body must be contained in the Session::_response_body attribute.

Removal of some debug logging that could consume a huge amount of memory in case
of bulk call.