WIP ADD FBUFFER FILTER
The job is not done yet.

The goal of fbuffer filter is to bufferize inputs for other
filters linked as outputs in fbuffer.conf.example.

At the beginning, the filter is a copy of TAnomaly and has
been modified on several points as described below:

_ Generator.cpp/hpp:
The configuration file parsing has been changed to
configure input_format and outputs. The way to configure the
buffer will be described in the documentation soon.

_ BufferThreadManager.cpp/hpp:
BufferThreadManager has changed a lot. TAomaly's version were
about managing one unique thread. But fbuffer needs several
threads (one per output). There is now a new additionnal class
called BufferThread which looks like TAnomaly's ThreadManager
because it manages one thread, where BufferThreadManaher handles
several BufferThread.
BufferThread class will every _interval seconds, query their own
redis_list to pick-up, if there's enough, the items in it to send
them to the correct filter.

_ BufferTask.cpp/hpp:
BufferTask is a class instanciated by the Generator and that
catches inputs and split them into the corrects Connectors
(described in AConnector.cpp/hpp and Connectors.hpp).
If a value needed by a linked output filter is missing, the
output is shut down and will never receive any data anymore.

_ AConnector.cpp/hpp:
Mother class for all the f<FilterName>Connector classes.
It defines the way a data is parsed from the raw input to
end up in the correct redis list (given in the
fbuffer.conf.example file).
All the sons of this class are defining what data should be sent
to the filter.

_Connectors.hpp:
A header file to prevent multiple includes in various files. It
contains all the f<FilterName>Connector.hpp so you only have to
include it.