WIP ADD FBUFFER FILTER
The fbuffer filter now read every {interval} (given in config) seconds
the {redis_list} REDIS list to pick up items if there are more than
{required_log_lines}. All 3 variables given in the config file.

AConnector.cpp/hpp and children:
Added the attribute _minLogLen and corresponding getter.
It stores the required_log_lines in the config file.

BufferThreadManager.cpp/hpp:
Uncommented in BufferThread REDISPopLogs and adapted to fit with the
multiple redis lists to handle.

Generator.cpp/hpp:
Changed the config parsing to read {required_log_lines} from it. And
send it to the corresponding connector.

My work is not tested, not cleaned, not documented, not finished yet.