WIP: CREATING A MORE UNIFORM ALERT FORMAT
**BREAKING**: The existing log format is replaced by a more uniform one.
The goal is to make it easier to read, parse, query and display.

The alert formating is no longer hard coded in the filters
to avoid mistakes.

The ability to add tags to the alert raised was implemented.
By default  the filters might contains some MITRE tags
indicating plausible threats type.
It can be configured through the "alert_tags" optional string list.

Implemented in Hostlookup for now.

TODO: All the other filters