ADD RSYSLOG LOOKUP TABLE FORMAT IN FHOSTLOOKUP
Added the possibility to use rsyslog lookup tables
as a hostlookup database.

By default if the loookup table contains string values, the score is 100
and the string is put in the alert "details.description" field.
If the value is an int, it is applied as the score with no description.
This last behavior is the default for text and json databases.

Added -o2 optimization flag to non debug compilation.