FIX THREAD CREATION ERROR UBUNTU
AThread _thread creation was made in AThread's constructor, which is
not a problem on HBSD, has an undefined behavior on Ubuntu (clang for
HBSD, gcc for Ubuntu).

Now AThread has a InitiateThread method that takes care of the _thread
initialisation and must be called after constructor (AThreadManager's
Start call in ThreadStart method). InitiateThread starts the actual
std::thread.